requirements.txt
Django==5.1.4
django-rest-framework==0.1.0